refactor(tree): move sequence utils that are only used for tests into test area
