+++ packages/dds/tree/src/feature-libraries/sequence-field/index.ts
@@ -42,11 +42,8 @@ export { NodeChangeRebaser, rebase } from "./rebase.js";
 export { invert, NodeChangeInverter } from "./invert.js";
 export { amendCompose, compose, NodeChangeComposer } from "./compose.js";
 export {
-	areComposable,
-	areRebasable,
 	getInputLength,
 	isDetach,
-	DetachedNodeTracker,
 	newCrossFieldTable,
 	newMoveEffectTable,
 	CrossFieldTable,
